# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/maquinas": {
      "post": {
        "operationId": "MaquinaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaquinaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Máquina criada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar uma nova máquina",
        "tags": [
          "maquinas"
        ]
      },
      "get": {
        "operationId": "MaquinaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de máquinas retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todas as máquinas",
        "tags": [
          "maquinas"
        ]
      }
    },
    "/maquinas/{id}": {
      "get": {
        "operationId": "MaquinaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da máquina",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da máquina retornados com sucesso."
          },
          "404": {
            "description": "Máquina não encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de uma máquina por ID",
        "tags": [
          "maquinas"
        ]
      },
      "patch": {
        "operationId": "MaquinaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da máquina",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaquinaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Máquina atualizada com sucesso."
          },
          "404": {
            "description": "Máquina não encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar uma máquina por ID",
        "tags": [
          "maquinas"
        ]
      },
      "delete": {
        "operationId": "MaquinaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da máquina",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Máquina removida com sucesso."
          },
          "404": {
            "description": "Máquina não encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover uma máquina por ID",
        "tags": [
          "maquinas"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS API",
    "description": "API documentation for NestJS with Swagger, Orval, and MikroORM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateMaquinaDto": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "description": "Código da máquina",
            "example": "MAQ-001"
          },
          "nome": {
            "type": "string",
            "description": "Nome da máquina",
            "example": "Torno CNC"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo da máquina",
            "example": "Torno"
          },
          "modelo": {
            "type": "string",
            "description": "Modelo da máquina",
            "example": "CNC-5000"
          },
          "fabricante": {
            "type": "string",
            "description": "Fabricante da máquina",
            "example": "Romi"
          },
          "capacidade": {
            "type": "string",
            "description": "Capacidade da máquina",
            "example": "500kg"
          }
        },
        "required": [
          "codigo",
          "nome",
          "tipo",
          "modelo",
          "fabricante"
        ]
      },
      "UpdateMaquinaDto": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "description": "Código da máquina",
            "example": "MAQ-001"
          },
          "nome": {
            "type": "string",
            "description": "Nome da máquina",
            "example": "Torno CNC"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo da máquina",
            "example": "Torno"
          },
          "modelo": {
            "type": "string",
            "description": "Modelo da máquina",
            "example": "CNC-5000"
          },
          "fabricante": {
            "type": "string",
            "description": "Fabricante da máquina",
            "example": "Romi"
          },
          "capacidade": {
            "type": "string",
            "description": "Capacidade da máquina",
            "example": "500kg"
          }
        }
      }
    }
  }
}